Allow the colour of features to be determined by a tag.
Allow the colour of features to be determined by a tag,
diagonal:colour. If the tag value is a hex colour, it's
used directly. If it's a floating point number, it's
clamped to the range [0,1) and used to determine a point
on a predefined gradient.
In passing, correct a bug in MutableTagsOverlayWorld that
dropped the value for unmodified tags, and add tests.