Add a progress bar widget in iPython.
Add a progress bar widget in iPython, which uses a JSON endpoint to poll
for updates. The involves changing the streaming GRPC interface - rather
than incrementally streaming progress, it sends a single tracker ID before
the result. Also add a new binary - scaffold - that makes it easy to play
with HTML fragments, like the widget.