Add snap_area_edges, flattern and points.
Add snap_area_edges, which returns a new area by snapped the edges of
an area to paths matching a given query, intended to be useful for
conflating boundary geometries to road networks.
Also add the useful remnants of a failed approach snapping at the
point level, specifically flattern, which flatterns collections,
and points, which extracts points from geometries.
Collections are now typed on the wire, allowing the type of empty
collections to be determined. Consequentely, the definition of Type
is begining to move into protocol buffers.