Make namespaces strings instead of integers.
Make namespaces strings, instead of a set of hardcoded integers. While
less efficient, it allows users to integrate arbitrary datasets under
a namespace of their choice, without predefined restrictions.
Namespaces are based on URLs related to the relevant data source.
This change allows us to remove the "user space" implementation of
namespaces in a5.
Note that this change breaks index formats, and so we bump the version
number.