Use custom types to modify the prototype Atlas UI.
Use custom types to modify the prototype Atlas UI. Functions return
a modification to the UI in the same way we handle modifications to
the world. This involved splitting Evaluate from EvaluateAsProto and
EvaluateAsValue, since we need both versions.
Remove dead code leftover from the namespace refactoring, and handle
the new string namespaces correctly in the Python library.