Autogenerate the Python API, moving from diagonal.a5 to diagonal_b6.
Output a JSON representation of functions available in B6 using
introspection in Go, and use that to autogenerate the corresponding
client library in Python.
Some function signatures change slightly, to aid autogeneration - for
example, Traversal is removed from graph traversal functions, as it
is no longer needed.
Remove the older version of the API, rather than backport the function
signature changes. This breaks code under experimental/ using the older
version of the API.