Added a coding exercise for SWE candiates.
Added a coding exercise for SWE candidares, that involves a simplified
accessibility analysis of the area around Granary Square with and without
the bridge to Somers Town. Includes golang code to generate the CSV data
sent to candidates, and a sample solution in Python using networkx.
As part of data generation, modify ingest-uprn to accept a b6 shell
expression to filter data.