Generate release artifacts inside a docker container.
Rework the Dockerfile to build the go binaries, Python package, and
Camden data within docker, before generating an image that just
includes the artifcats, and not the build tools. Also add a
placeholder README and an LICENSE.
In passing, add the same options for disk-based index build to
b6-connect that exist in b6-ingest-osm.