Show the version in the UI, and verify it in GRPC requests.
Stamp the current version number into the b6 backend binary,
and show it in the web UI. Also include it in the generated
Python client code, and send it along with GRPC requests.
Fail requests that come from a Python library with a different
major version number. Also use the version number in the PIP
project description, and the binary tarball in the docker build.