Fix a sightline panic, and allow features to be passed back in Python.
Fix a sightline panic that occured when an edge ended exactly on the
reference line between the point and the North pole. As it was part
of the same bug report, also allow features to be passed back to the b6
server by encoding them as a call to find-feature with their ID.
In passing, correct a bug that prevented str() from working on features
given we now use string namespaces, not integers.

gh-42: str() on features, and sending features to the backend
gh-43: sightline crash