Add support for collection literals in the b6 shell and Python.
Add supprt for collection literals in the b6 shell and Python. In the
shell, you can say:
{"primary", "secondary"}
for a collection with implicit integer keys from the value's index in
the literal, or:
{highway="motorway": 2, highway="primary": 6}
for a collection with explicit keys and values (here tags and integers)

Similarly in Python, lists and dicts are converted into collections. Keys
and values are evaluated server side, and hence can be arbitrary b6
expressions, not just literals.

Fixes #66