Support map() with functions that return a literal.
Support map() with functions that return a literal. Previously, we
only supported b6 expressions, eg:
b6.map(buildings, lambda: building: building.area())
as we relied on inferring the type of the collection from the expression.
To support literals, we build a map of literal types to the corresponding
expression result type in the generated code.