Support the use of queries as functions in Python.
Support the use of queries as functions in Python, as they're implictly
converted to a function taking a feature and returning a bool on the
server side. Add a new matches() function, to allow the conversion to be
made explicit when desired. Deprected has_key(), in favour of using the
query returned by keyed().