Support expressions as features. (#131)
Support expressions as features within the world. Adds a new, non proto,
based representation for expressions, which will become the standard internal
representation for, eg, the VM in a future CL. Support exporting expressions
as YAML, though not yet in a compact world. Moves Tags from b6/ingest to b6/
for wider reuse.