Add ExpressionFeature. (#132)
Add ExpressionFeature, to support saving expressions into the world, as a
first step to supporting clustering analysis results in the UI. Introduces
b6.Expression, a non-proto expression representation suitable for the World
interface, and uses it in the VM. Moves the Collection interface from api to
b6, and takes the opportunity to make it type safe using generics.

Includes a couple of minor breaking API changes, reflected in the Python tests:
s2 functions now string a collection if string tokens with int index keys
(previously, the returned collections with the token used for both the key and
the value, which isn't a pattern we use elsewhere). Also, applying a change
to the world used to return an ingest.AppliedChange, dating from when we
didn't support collections. Now it returns a Collection[FeatureID, FeatureID].