Make materialise() add collections and expressions in the world.
Change the interface of materialse(), so that it now takes a function
that returns a collection, calls that function, and then adds both
that collection, and the function that created it to the world.

Includes the Python API for CollectionFeature and ExpressionFeature,
which necessitated merging feature.proto and api.proto, as they now
depend on each other.