Serve the vite+react frontend from b6.
Serve the vite+react frontend from the b6 backend binary, integrating it
into the continuous build. The new (which I'm naming v2) frontend bundle
is served alongside the existing frontend, via a new /v2.html
entrypoint.

In passing, correct the error returned for missing files from
MergedFilesystem, so that the server returns a 404, and not 500.