Support separate worlds, keyed by the root feature ID.
Support separate worlds, using the ID that's currently used for the feature
that configures the UI to select which world. Support passing this ID
through Python, and add it to the tile requests from javascript.
TODO: this ID is currently called "context", "root", or "UI context", we
should standardise on something.