Add a function to materialise the result of a map operation. (#182)
Add a function to materialise the result of a map operation. We call
a given function for every feature in a collection, materialising the
result of that function as a feature with a randomly assigned ID, before
adding a new collection feature with the given ID that references those
results.
This will enable us to materialise the results of, for example, accessibility
computations per building, rather than as one large collection for a wide
area.