Render histograms by creating a collection in the world. (#181)
Render histograms by creating a collection in the world, and passing
the ID of that collection through to the tile renderer, instead of
including the IDs of bucketed features in the JSON response. This makes
the size of the response tractable. These features are currently leaked,
but that'll be addressed in the future (by deleting features when the
relevant stack is closed in the UI).

Introduces a new method on CollectionFeature to find the first value for
a given key, which can make use of binary search if the collection is
sorted (otherwise tile rendering becomes very slow with large collections).

In passing, make the tests for histogram rendering non-flakey by explictly
ordering keys as they're encountered.