Support the exist tag serialisation format for tags with string values. (#185)
We currently use tags with string values in configuration files for, eg,
the basemap rendering rules. Support the old way of serialising them to
YAML (as, eg, #highway=pedestrian), while also supporting the serialisation
of more complex tag values via the existing YAML serialisation code for
literals in expressions (which are the same thing - so b6.Literal and
b6.Value should merge).